+++ CHANGELOG.md
@@ -7,7 +7,7 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 ### Added
 ### Removed
 ### Changed
-- The utility function `dhtSync` parameter `timeoutMs` now defaults to 60000 milliseconds.
+- The utility function `dhtSync` parameter `timeoutMs` now defaults to 60000 milliseconds, and the parameter `intervalMs` now defaults to 500 milliseconds.
 
 ### Fixed
 - Return an error if spawning `kitsune2-bootstrap-srv` fails